% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{convert.size}
\alias{convert.size}
\title{Convert a 2D size from one unit to another.}
\usage{
convert.size(dim, from = "inches", to = "user")
}
\description{
Convert a 2D size from one unit to another.
}

