% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancyplots.R
\docType{package}
\name{fancyplots}
\alias{fancyplots}
\title{Scatter Plots with Logarithmically Scaled X Axis!}
\description{
This package makes some very fancy ggplot scatter plots with the x axis logarithmically scaled
}
\note{
This package was created for STAT 545B at the University of British Columbia.
}
\author{
Kaitlin Sullivan
}
