% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{linear.heatmap}
\alias{linear.heatmap}
\title{Plot values as linear heatmaps.}
\usage{
linear.heatmap(x, center = 0.5, minval = 0, maxval = 1,
  palettes = brewer.pal(9, "Reds"), nslices = 50, xlab = "",
  spacing = 0.05, bottom.to.top = TRUE, ...)
}
\arguments{
\item{x}{vector of numeric values}

\item{center}{central value; each value of x will be shown relative to its difference from `center`}

\item{palettes}{palettes for chosing colors for values less than and greater than `center. Either a vector or
a list of two vectors, where each vector contains two or more colors, ordered from closest to furthest from
`center`.}

\item{nslices}{number of color slices to use in each gradient}

\item{xlab}{x-axis label}

\item{spacing}{amount of space [0-0.5) between each heatmap}

\item{bottom.to.top}{if TRUE, plot values from the bottom to the top of the plot, otherwise top-down}

\item{...}{additional arguments to be passed to `plot`}

\item{min}{minimum value; must be <= `center`}

\item{max}{value; must be >= `center`}
}
\description{
Plot values as linear heatmaps.
}
\examples{
vals <- runif(10)
names(vals) <- paste('Sample', 1:10)
par(mar=c(4,5,1,1))
linear.heatmap(vals, xlab='Allelic Ratio (B/A)')
}

