% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{make.cb.pal}
\alias{make.cb.pal}
\title{Create an arbitrary-sized palette from an RColorBrewer palette.}
\usage{
make.cb.pal(name, cb.colors = 8, pal.colors = 255)
}
\arguments{
\item{name}{RColorBrewer palette name}

\item{cb.colors}{number of colors to use from the RColorBrewer palette}

\item{pal.colors}{number of colors to generate from the RColorBrewer palette}
}
\description{
Create an arbitrary-sized palette from an RColorBrewer palette.
}

