% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{make.ramp}
\alias{make.ramp}
\title{Generate a color ramp function for a given pallate and intensity range.}
\usage{
make.ramp(pal, start = 1, end = 100)
}
\description{
Generate a color ramp function for a given pallate and intensity range.
}

