% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{manhattan.bimbam}
\alias{manhattan.bimbam}
\title{Plot results of BIMBAM run. Draws manhattan plots based on BF. Expects
at least 4 column data frame: pos, chr, bf, pv.}
\usage{
manhattan.bimbam(bb, outdir, main = "", outfile = "bayes_manhattan.pdf",
  prior = NULL, threshold = c(0.99, 0.95), as.prob = FALSE, ...)
}
\description{
Plot results of BIMBAM run. Draws manhattan plots based on BF. Expects
at least 4 column data frame: pos, chr, bf, pv.
}

