% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{manhattan.freq}
\alias{manhattan.freq}
\title{Manhattan plot for results from frequentist analysis. Requires at least 3 column data frame:
CHR, BP, and P, corresponding to chromosome number, genomic coordinate, and p-value. Missing
values (where regression failed to converge, for instance) should be recoded NA before reading
into R. The plot alternates between two colors.}
\usage{
manhattan.freq(d, mapping = NULL, colors = c("steelblue", "gray50"),
  chrms = 1:19, threshold = c(1e-05, 1e-08),
  sub = "Frequentist Association", ...)
}
\description{
Manhattan plot for results from frequentist analysis. Requires at least 3 column data frame:
CHR, BP, and P, corresponding to chromosome number, genomic coordinate, and p-value. Missing
values (where regression failed to converge, for instance) should be recoded NA before reading
into R. The plot alternates between two colors.
}
\note{
Copied and modified from Stephen Turner (http://GettingGeneticsDone.blogspot.com/).
}

