% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{manhattan}
\alias{manhattan}
\title{Generic Manhattan plotting function. Requires a data frame with at least 3 columns:
CHR, BP and STAT.}
\usage{
manhattan(d, colors, ylab, ymax = NULL, cex.x.axis = 1, annotate = NULL,
  anno.col = "green", main = NULL, sub = NULL, threshold = NULL,
  lty.threshold = 1, col.threshold = "red", dev.type = c("jpg", "pdf"),
  dev.file = NULL, width = 20, height = 6, ...)
}
\description{
Generic Manhattan plotting function. Requires a data frame with at least 3 columns:
CHR, BP and STAT.
}

