% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gwas.R
\name{plot.effects}
\alias{plot.effects}
\title{Effect size plotting function. Requires a data frame with at least 3 columns: CHR, BP and BETA.}
\usage{
plot.effects(d, chrm = NULL, rng = NULL, colors = c("steelblue",
  "gray50"), ylab = "Effect Size", ylim = NULL, cex.x.axis = 1,
  annotate = NULL, main = NULL, sub = NULL, pdf.file = NULL,
  width = 20, height = 6, ...)
}
\description{
Effect size plotting function. Requires a data frame with at least 3 columns: CHR, BP and BETA.
}

