% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot.matrix}
\alias{plot.matrix}
\title{Plot a matrix as a heatmap.}
\usage{
plot.matrix(x, xlab = "", ylab = "", xLabels = NULL, yLabels = NULL,
  grid.on = TRUE, cex.axis = 0.7, grid.col = "white", grid.lwd = 1,
  ColorRamp = NULL, reverse.y = TRUE, zlim = NULL, pdf.file = NULL,
  pdf.width = NULL, pdf.height = NULL, region = c("all", "upper.tri",
  "lower.tri"), axis.ivl = 1, hide.x.names = FALSE, hide.y.names = FALSE,
  col.dendro = NULL, ...)
}
\description{
Plot a matrix as a heatmap.
}

