% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabular.R
\name{plot.table}
\alias{plot.table}
\title{Print a table as a figure using grid graphics.}
\usage{
plot.table(data, plotters = list("c"), col.widths = NULL,
  table.border = FALSE, table.border.gpar = NULL, cell.mar = 0.05,
  cell.border = FALSE, cell.border.gpar = NULL,
  header.gpar = gpar(fontface = "bold"), default.gpar = gpar())
}
\arguments{
\item{data}{matrix or data frame}

\item{plotters}{list of plotters to use (will be recycled if length is shorter than
`ncol(data)`). Each element is either a name of a pre-defined plot function, or a
list with `fn` set to the plotter function name and additional elements being
arguments that will be passed to the plotting function.}

\item{col.widths}{vector of column widths}

\item{table.border}{logical; whether to draw a border around the table.}

\item{table.border.gpar}{list of graphical parameters to use when drawing borders.}

\item{cell.mar}{size of margin around values in table cells.}

\item{cell.border}{logical; whether to draw borders around cells.}

\item{cell.border.gpar}{list of graphical parameters to use when drawing cell borders.}

\item{header.gpar}{list of graphical parameters to use when drawing column headers.}

\item{default.gpar}{list of default graphical parameters to use when drawing cell values.}
}
\description{
Print a table as a figure using grid graphics.
}
\note{
If you are using knitr, it is much better to use pander to format tables.
}

