% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/venn.R
\name{plot.venn}
\alias{plot.venn}
\title{Draw a Venn diagram of intersections between lists of items. Supports between 2-5 lists.}
\usage{
plot.venn(lists, plotData = TRUE, returnFig = FALSE, pal = c("lightgreen",
  "orange", "violet", "yellow", "brown"), ...)
}
\arguments{
\item{lists}{matrix with between 2-5 columns or list with 2-5 vectors.}

\item{plotData}{logical whether to draw the plot}

\item{returnFig}{logical whether to return the figure object}

\item{pal}{vector of colors to use for filling circles

TODO: allow for different colors in each region}
}
\description{
Draw a Venn diagram of intersections between lists of items. Supports between 2-5 lists.
}

