% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ape_phyo.R
\name{plottree.phylo}
\alias{plottree.phylo}
\title{Convenience function to plot an ADE phylo.}
\usage{
plottree.phylo(p, type = "phylogram", names.file = NA, pdf.file = NULL,
  pdf.width = 20, pdf.height = 20, draw.branch.labels = FALSE,
  convert.branch.labels = NULL, ...)
}
\description{
Convenience function to plot an ADE phylo.
}

