% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ade4_phylo.R
\name{radial.phylog}
\alias{radial.phylog}
\title{Draw a radial tree for an ade4 phylog object.}
\usage{
radial.phylog(phylog, circle = 1, show.leaves = TRUE,
  show.leaf.nodes = TRUE, labels.leaves = names(phylog$leaves),
  cex.leaves = 1, bg.leaves = "black", pch.leaves = 21,
  fg.text.leaves = "black", ragged.leaves = FALSE, pad.leaves = 0,
  cex.leaf.points = 1, number.leaves = FALSE, cex.leaf.numbers = 0.5,
  labelptr.show = TRUE, labelptr.col = grey(0.7), labelptr.pad = NULL,
  show.nodes = FALSE, labels.nodes = phylog$nodes, cex.nodes = 1,
  bg.nodes = "white", pch.nodes = 21, bg.colramp.nodes = NULL,
  fg.nodes = "black", adj.nodes = 0, lwd.branch = 1,
  lwd.branch.default = 1, col.branch = "black",
  col.branch.default = "black", col.branch.colramp = NULL,
  weights.branch = NULL, col.center = "red", cex.center = 2,
  legend.nodes = FALSE, legend.title.nodes = "Nodes",
  legend.branch = FALSE, legend.title.branch = "Branches", main = NULL,
  pdf.file = NULL, pdf.width = 7, pdf.height = 7)
}
\description{
Draw a radial tree for an ade4 phylog object.
}

