% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{rgb.alpha}
\alias{rgb.alpha}
\title{Modifiy colors (names or rgb vectors) with an alpha value.}
\usage{
rgb.alpha(r, alpha = 1)
}
\description{
Modifiy colors (names or rgb vectors) with an alpha value.
}

