% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot_logx.R
\name{scatter_plot_logx}
\alias{scatter_plot_logx}
\title{scatter_plot_logx: plotting a scatter plot with a logarithmically scaled x axis.}
\usage{
scatter_plot_logx(data, x, y, logx = TRUE, verbose = TRUE, na.omit = TRUE)
}
\arguments{
\item{data}{Dataframe or tibble.}

\item{x}{String containing the name of a column from data to be the x axis. All column values \strong{must} be \strong{\emph{numeric}}.}

\item{y}{String containing the name of a column from data to be the y axis.}

\item{logx}{A logical (default = TRUE.)}

\item{verbose}{A logical (default = TRUE.)}

\item{na.omit}{A logical (default = TRUE.)}
}
\value{
A ggplot scatter plot with the x axis logarithmically scaled.
}
\description{
scatter_plot_logx: plotting a scatter plot with a logarithmically scaled x axis.
}
\examples{
\dontrun{scatter_plot_logx(gapminder, "pop", "lifeExp")}
}
