% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex.triangle}
\alias{simplex.triangle}
\title{A simplex is a way to plot a specific type of multi-variate information in two-dimensions.
Each data point is taken to be a list of proportions that sum to 1.0. An example of data
that can be plotted with a simplex is admixture assignments from three or more founder
populations: the higher the value of a specific variable (population), the more of the
sample's admixture is derived from that population.}
\usage{
simplex.triangle(series, main = NULL, corner.labels = NULL, cex = NULL,
  show.ticks = FALSE, show.legend = FALSE, legend.pos = "topleft",
  plotwidth = 7, new.plot = TRUE)
}
\arguments{
\item{series}{either a list or a list of lists (see details).}

\item{main}{plot title.}

\item{corner.labels}{labels for each of the vertices.}

\item{cex}{point cex.}

\item{show.ticks}{show tick marks on the sides of the triangle.}

\item{show.legend}{whether to show a legend defining the colors used for each variable.}

\item{legend.pos}{where to put the legend.}

\item{plotwidth}{size of plot (in inches).}

\item{new.plot}{logical whether to create a new plot or overplot on the existing device.}
}
\description{
Points are plotted in the interior of a shape with a number of vertices equal to the number
of variables (i.e. a triangle for data points in three dimensions, a square for four dimensions,
etc). The higher the value of a variable, the closer the point is to the vertex representing
that variable. The color of the point is also determined by it's closeness to each vertex.
}
\details{
If `series` is a list, then each element is a Nx3 matrix, where each row is translated
into a data point. The sum of each row in the data matrix must be 1. If `series` is a
list of lists, each element has a `data` member, which is the matrix, and must also
have pch and colramp elements. If `show.legend` is TRUE, each list must also have a name
element. Other optional elements are: highlight.idx and highlight.col, where highlight.idx
can be an index or list of indices to highight with the specified color (this works best if
the colramp is grayscale).
}

